requirements.txt
python-dotenv
pandas
numpy
pymongo
certif
pymongo[srv]==3.10
pymongo
scikit-learn

##-e .
